Remove footnote
Line 39 (id=sc15) is not part of the text, but is a note by the translator. I can clearly see it in the printed version, it's a footnote. We don't usually show such notes, so it should probably be deleted. If that is the right thing to do, please merge this pull request, otherwise just close it.