Update dn1.html
This fixes the link to the Buddhist Publication Society from a strange link to the book cover to a link to the BPS homepage. 

I believe this is better than any direct link on their website because from time to time they reorganize the website so the links are fragile.